Task 4: add loggers
