+++ .eslintrc.json
@@ -42,6 +42,8 @@
     "dot-notation": "off",
     "node/no-missing-import": ["error", {
       "tryExtensions": [".js", ".ts"]
-    }]
+    }],
+    "@typescript-eslint/no-unused-vars": ["warn", { "argsIgnorePattern": "^_" }],
+    "no-unused-vars": ["warn", { "argsIgnorePattern": "^_" }]
   }
 }
