+++ src/middleware/logger/unhandledRejections.logger.ts
@@ -1,17 +1,7 @@
-import winston from 'winston';
-import expressWinston from 'express-winston';
+import fs from "fs";
 
-export const errorsLogger = expressWinston.errorLogger({
-    transports: [
-        new winston.transports.Console(),
-        new winston.transports.File({
-            filename: 'error.log',
-            level: 'error',
-            format: winston.format.combine(
-                winston.format.uncolorize(),
-                winston.format.json()
-            )
-        }),
-    ],
-    msg: "HTTP {{req.method}} {{req.url}} | query parameters: {{JSON.stringify(req.query)}} | body: {{JSON.stringify(req.body)}} | status code: {{res.statusCode}}", // optional: customize the default logging message. E.g. "{{res.statusCode}} {{req.method}} {{res.responseTime}}ms {{req.url}}"
-});
+export const unhandledRejectionsLogger = (reason: {message: string }) => {
+    console.error(`Unhandled rejection detected: ${reason.message}`);
+    fs.writeFileSync('./error.log', `Unhandled rejection detected: ${reason.message}`, {flag:'a'})
+    process.exit(1);
+}
